Added: fungi phylogeny list
This is an example table of the taxonomy file needed to run the script.  If you want the .dmnd binary database used in the paper describing this work, contact me and I will share it with you. This taxonomy file matches that database. Alternatively, you could construct your own, incorporating newer fungal genomes.